Optimod is now functional
Hook controllers are finished, but configuration saving might need a bit more work. ConVars are also going to need a button to set the convar to the optimal value, as there is nothing that tells you what the optimal values is aside from the effectivity color. Put in a request for more German translations for the new text needed, and changed some of the localization keys.